requirements.txt
opencv_python==4.5.5.62
Pillow==8.3.1
torch==1.10.2
torchvision==0.11.3
